+++ extensions-builtin/forge_legacy_preprocessors/annotator/oneformer/detectron2/evaluation/pascal_voc_evaluation.py
@@ -294,7 +294,7 @@ def voc_eval(detpath, annopath, imagesetfile, classname, ovthresh=0.5, use_07_me
     rec = tp / float(npos)
     # avoid divide by zero in case the first detection matches a difficult
     # ground truth
-    prec = tp / np.maximum(tp + fp, np.finfo(np.float64).eps)
+    prec = tp / np.maximum(tp + fp, np.finfo(np.float32).eps)
     ap = voc_ap(rec, prec, use_07_metric)
 
     return rec, prec, ap
